Depend on psycopg2-binary
This is needed for local development, as compiling `psycopg2` (doesn't ship binaries) requires `pg_config`. Which is shipped as part of at least the `postgresql` Homebrew formula, and likely to be absent in all cases in which this library can be used:

- When using this library directly, as users are instructed to run PostgreSQL inside Docker.
- When using software that depends on this library, as an RDBMS may not even be used in any capacity.

As `dh_python3` does not contain a mapping for `psycopg2-binary` (the module name is not `psycopg2`, as is the case with the regular `psycopg2`, but `psycopg2_binary`), and `python3-psycopg2` does not ship a `pydist` file, the dependency must be mapped manually using a `py3dist-overrides` file.

A disadvantage of this approach: the `psycopg` project's official documentation discourages using `psycopg2-binary` in production environments, as it contains statically linked dependencies (such as `libssl`). Therefore, we would like to depend on `psycopg2` (source package) for production environments, and on the binary version in local contexts only. However, this is not feasible, as Python doesn't support dropping dependencies for certain 'extras', see: https://discuss.python.org/t/optional-dependency-groups-omitting-package-requirements/20833. The only way would be to ship multiple packages (e.g. a new `python3-cyberfusion-database-support-pgsql-binary`), but that would escalate very quickly e.g. when adding another RDBMS with a similar situation. And as we use Debian packages in production (depending on the production-ready `python3-psycopg2` Debian package), this is not relevant for us.